Split out arrow-schema (#2594)
