+++ Cargo.toml
@@ -18,6 +18,7 @@
 [workspace]
 members = [
         "arrow",
+        "arrow-schema",
         "parquet",
         "parquet_derive",
         "parquet_derive_test",
